+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+* `Animation` struct with `from_frame_duration`, `update` and `current_frame` methods
 * Support `no_std` by disabling `std` feature
 
 [Unreleased]: https://github.com/jcornaz/franim/compare/...HEAD